move purity function to states.measures
