+++ qiskit/quantum_info/__init__.py
@@ -59,6 +59,7 @@ Measures
    :toctree: ../stubs/
 
    state_fidelity
+   purity
    average_gate_fidelity
    process_fidelity
    gate_error
@@ -109,8 +110,8 @@ from .operators.measures import process_fidelity
 from .operators import average_gate_fidelity
 from .operators import gate_error
 from .states import Statevector, DensityMatrix
-from .states import state_fidelity, partial_trace
-from .states.states import basis_state, projector, purity
+from .states import state_fidelity, partial_trace, purity
+from .states.states import basis_state, projector
 from .random import random_unitary, random_state, random_density_matrix
 from .synthesis import (TwoQubitBasisDecomposer, euler_angles_1q,
                         two_qubit_cnot_decompose)
