+++ qiskit/quantum_info/states/states.py
@@ -57,17 +57,3 @@ def projector(state, flatten=False):
     if flatten:
         return density_matrix.flatten(order='F')
     return density_matrix
-
-
-def purity(state):
-    """Calculate the purity of a quantum state.
-
-    Args:
-        state (ndarray): a quantum state
-    Returns:
-        float: purity.
-    """
-    rho = np.array(state)
-    if rho.ndim == 1:
-        return 1.0
-    return np.real(np.trace(rho.dot(rho)))
